+++ app/src/main/java/com/pbs/expenseApp/ui/screens/user/UserEntryViewModel.kt
@@ -4,8 +4,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.pbs.expenseApp.ui.database.entities.User
 import com.pbs.expenseApp.ui.database.repositories.UserRepository
+import kotlinx.coroutines.async
+import kotlinx.coroutines.launch
 
 class UserEntryViewModel(private val userRepository: UserRepository): ViewModel() {
     var userUiState by mutableStateOf(UserUiState())
@@ -17,8 +20,18 @@ class UserEntryViewModel(private val userRepository: UserRepository): ViewModel(
         )
     }
 
+    suspend fun isUserExists(uuid: String): Boolean {
+        var userExists = false
+        viewModelScope.async {
+            userExists = userRepository.isUserExists(uuid)
+        }.await()
+        return userExists
+    }
+
     suspend fun saveUser(uuid: String) {
-        userRepository.insertUser(userUiState.userDetails.toUser(uuid))
+        viewModelScope.launch {
+            userRepository.insertUser(userUiState.userDetails.toUser(uuid))
+        }
     }
 }
 
@@ -36,7 +49,7 @@ fun UserDetails.toUser(uuid: String): User = User(
     monthlySavings = monthlySavings.toDoubleOrNull() ?: 0.0,
 )
 
-fun User.toUserUiState(isEntryValid: Boolean = false): UserUiState = UserUiState(
+fun User.toUserUiState(): UserUiState = UserUiState(
     userDetails = this.toUserDetails()
 )